+++ tests/formatter/input-output-pairs/assert.out.gd
@@ -1,3 +1,3 @@
 func foo():
-	assert 1 == 1
 	assert(1 == 1)
+	assert(1 == 1, "asdf")